Make $order_id a required parameter
Since `$order_id` is now necessary for retrieving the gateway specific settings, we should make it its own parameter to prevent forgetting to include it.

To prevent breaking existing behavior, it is used as a 2nd parameter. Making it a 2nd argument will also avoid what is actually the arguments array to be passed as the order ID.